refactor: isolate pydantic ai 1.x patch setup
